==> notebooks/0. Prerequisites.ipynb <==
# Prerequisites
This notebook installs the required dependencies and downloads the data needed to perform the analysis and mdoel training.

## Install dependencies

# Install required dependencies
!pip install -r requirements.txt

## Download dataset and helper files

# Download dataset and helper files into root folder
# https://console.cloud.google.com/storage/browser/...

# Download Quick Draw! .npy files from Google Cloud into the quickdraw folder
# https://console.cloud.google.com/storage/browser/quickdraw_dataset/full/numpy_bitmap
