==> notebooks/6.3. QD Transfer.ipynb <==
import sys
import json
from pathlib import Path

from PIL import Image
from tqdm import tqdm
import numpy as np
import pandas as pd

from bananas.utils import images
from bananas.dataset import DataSet, DataType, Feature

# Root path of project relative to this notebook
ROOT = Path('..')

sys.path.insert(1, str(ROOT / 'scripts'))
from datamodels import *
from utils import *

### Load QuickDraw data

X, y = [], []
root = ROOT / 'quickdraw'
for file in tqdm(list(root.iterdir())[1:5]):
    name = file.name.split('.')[0]
    for img in tqdm(np.load(file)[:300], leave=False):
        y.append(name)
        img = img.reshape(28, 28)
        img = images.ndarray_to_pil(img)
        img = img.resize((300, 300))
        img = images.pil_to_ndarray(img)
        X.append(img.reshape((1, *img.shape)).astype(float) / 255)

### Build dataset and instantiate learner

ds = DataSet([Feature(X)], target=Feature(y), batch_size=32)

from bananas.statistics.scoring import ScoringFunction
from quick_draw_learner import QDClassifier
learner = QDClassifier(random_seed=0)

### Train model using Quick Draw data

import pickle

model_path = ROOT / 'models' / 'quick_draw_model.pth'
if model_path.exists():
    # If the model has already been trained, load it
    with open(model_path.absolute(), 'rb') as fh:
        learner.model_ = pickle.load(fh)

else:
    # Otherwise train it now
    learner.train(ds.input_fn, progress=True, max_score=.975, max_steps=1000)
    with open(model_path.absolute(), 'wb') as fh: pickle.dump(learner.model_, fh)

### Save pre-trained model to be used as starting point

quick_draw_learner = learner
quick_draw_model = learner.model_

### Read subject data from local file

df = pd.read_csv(ROOT / 'datasets' / 'subject_diagnosis.csv', index_col=0)

# Convert non-primitive fields and normalize paths
df['processed_path'] = df['processed_path'].apply(lambda x: ROOT / x)
df['image_path'] = df['image_path'].apply(lambda x: ROOT / x)
df['template_path'] = df['template_path'].apply(lambda x: ROOT / x)
df['drawing_box'] = df['drawing_box'].apply(lambda x: Box.load(x))
df['template_box'] = df['template_box'].apply(lambda x: Box.load(x))

# Remove all unnecessary columns from our dataset
feat_keys = ['processed_path']
group_columns = ['diagnosis']
df = df[group_columns + feat_keys].copy()

df.head()

### Create a file to hold intermediate results

results_file = ROOT / 'results' / 'qd_transfer.csv'

if not results_file.exists():
    pd.DataFrame(columns=[
        'Key',
        'Trial',
        'Subset splits', 
        'Batch size', 
        'Random seed', 
        'Accuracy',
        'Precision',
        'Recall',
        'Area under ROC']).set_index('Key').to_csv(results_file)

trial_results = pd.read_csv(results_file, index_col='Key')

### Train transfer learning model

from itertools import product, combinations
    
# Define all possible hyperparameters
batch_sizes = [24, 32, 48, 56, 64]
test_splits = [.2, .25]
validation_splits = [.2, .25]

# Initialize random number generator without seed to randomize hyperparamters
rnd = np.random.RandomState()

# Cross product all hyperparameters
parameter_combinations = list(product(
    batch_sizes, test_splits, validation_splits))
rnd.shuffle(parameter_combinations)

target_label = 'SANO'
target_column = 'diagnosis'

from bananas.core.mixins import HighDimensionalMixin
from bananas.sampling.cross_validation import DataSplit
from bananas.statistics import scoring
from coconuts.learners.transfer_learning import TransferLearningModel

# Perform 3 trials per parameter set to later compute the average
trial_count = 3

for batch_size, test_split, validation_split in tqdm(parameter_combinations, leave=False):
    
    trial_params = {
        'Subset splits': (test_split, validation_split),
        'Batch size': batch_size,
    }
    
    # Execute the independent trials
    for trial_num in tqdm(range(trial_count), leave=False, desc='Trial'):
        trial_params['Trial'] = trial_num
        
        # If these parameters have already been tried, we can skip trial
        trial_key = '|'.join(['%s=%s' % (k, str(v)) for k, v in trial_params.items()])
        if trial_key in trial_results.index: continue

        # Create a single feature containing all image data
        features = [Feature(
            ImageAugmenterLoader(df['processed_path'].values),
            kind=DataType.HIGH_DIMENSIOAL,
            sample_size=10,
            random_seed=0)]

        # Define target feature
        target_feature = Feature(
            (df[target_column] == target_label).values, random_seed=0)

        # Always re-initialize the randopm seed
        random_seed = trial_num

        while True:

            # Change seed at every iteration to ensure different dataset split
            random_seed = np.random.RandomState(seed=random_seed).randint(1E6)
            
            # Build dataset, making sure that we have a left-out validation subset
            dataset = DataSet(
                features,
                name=target_label,
                target=target_feature,
                random_seed=random_seed,
                batch_size=batch_size,
                test_split=test_split,
                validation_split=validation_split)

            # Compute test class balance to tell what minimum accuracy we should beat
            test_idx = dataset.sampler.subsamplers[DataSplit.VALIDATION].data
            test_classes = target_feature[test_idx]
            test_class_balance = sum(test_classes) / len(test_classes)

            # Rebuild dataset unless test class balance is within 5% of ground truth
            true_class_balance = sum(target_feature[:] / len(target_feature))
            if abs(test_class_balance - true_class_balance) < .05: break

        # Instantiate learner using pre-trained model
        learner = TransferLearningModel(quick_draw_learner, freeze_base_model=False)

        # Train learner using train dataset
        learner.train(dataset.input_fn, progress=True, max_steps=1000)

        # Test learner predictions using left-out validation dataset
        X, y = dataset[test_idx]
        y = learner.label_encoder_.transform(y)
        y_ = learner.predict_proba(X)

        # Store trial results
        trial_results.loc[trial_key] = {
            **trial_params,
            'Random seed': random_seed,
            'Accuracy': scoring.score_accuracy(y, y_),
            'Precision': scoring.score_precision(y, y_),
            'Recall': scoring.score_recall(y, y_),
            'Area under ROC': scoring.score_auroc(y, y_),
        }
        
        trial_results = trial_results.sort_values('Accuracy', ascending=False)
        trial_results.to_csv(results_file)

trial_results.head(15)
